requirements.txt
streamlit
jamaibase
python-dotenv
notion-client
openai
openai-whisper
opencv-python
pytesseract
yt-dlp
streamlit-authenticator
python-slugify
supabase